# api/User-openapi.yaml
openapi: 3.1.0
info:
  title: User Service Papich API
  description: "Документация API для микросервиса юзеров Папич. Управление юзерами\
    , ролями и сессиями"
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Generated server url
...
paths:
...
  /roles:
    post:
      tags:
        - Роли/должности
      summary: Создать роль
      description: "Присваивает юзеру роль. Если у текущего юзера (отправившего за\
        прос) такая же роль, то она передается: у текущего она удаляется, у указан\
        ного - появляется"
      operationId: setRoleToUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleSetDto"
        required: true
      responses:
        "201":
          description: Успешное создание (присваивание) роли
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleResponseDto"
        "403":
          description: Нет прав управлять ролью
        "404":
          description: Не найден юзер или роль
        "409":
          description: Юзер уже назначен на роль
        "507":
          description: Не удалось передать роль
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
        "401":
          description: Запрос не авторизован
      security:
        - bearerAuth: [ ]
...
  /roles/edit:
    patch:
      tags:
        - Роли/должности
      summary: Изменить роль
      description: Меняет юзеру роль
      operationId: editRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleEditDto"
        required: true
      responses:
        "200":
          description: Успешное изменение роли
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleResponseDto"
        "403":
          description: Нет прав управлять ролью
        "404":
          description: Роль не найдена
        "409":
          description: "Юзер уже назначен на роль, или кто-то внес изменения во вр\
            емя выполнения запроса"
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
        "401":
          description: Запрос не авторизован
      security:
        - bearerAuth: [ ]
...
  /users/get/with/roles/{id}:
    get:
      tags:
        - Юзеры
      summary: Получить инфу с ролями о юзере по ID
      description: Возвращает инфу с ролями о юзере по ID
      operationId: getUserWithRoles
      parameters:
        - name: id
          in: path
          description: ID юзера
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешное получение инфы о юзере
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseWithRoleDto"
        "401":
          description: Запрос не авторизован
        "404":
          description: Юзер не найден
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
...
  /roles/get/{id}:
    get:
      tags:
        - Роли/должности
      summary: Получить роль по ID
      description: Возвращает роль по ее ID
      operationId: getRole
      parameters:
        - name: id
          in: path
          description: ID роли
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешное получение роли
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleResponseDto"
        "404":
          description: Роль не найдена
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
        "401":
          description: Запрос не авторизован
      security:
        - bearerAuth: [ ]
  /roles/get/user/roles/{userId}:
    get:
      tags:
        - Роли/должности
      summary: Получить все роли (связки юзера и роли в БД) юзера по его ID
      description: Возвращает список ролей юзера по его ID
      operationId: getRoleByUserId
      parameters:
        - name: userId
          in: path
          description: ID юзера
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешное получение ролей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleResponseDto"
        "404":
          description: Юзер не найден
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
        "401":
          description: Запрос не авторизован
      security:
        - bearerAuth: [ ]
  /roles/get/user/roles/all/{userId}:
    get:
      tags:
        - Роли/должности
      summary: "Получить все роли (названия, в том числе дочерние) юзера по его ID"
      description: Возвращает список названий ролей юзера по его ID
      operationId: getAllRolesByUserId
      parameters:
        - name: userId
          in: path
          description: ID юзера
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешное получение ролей
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "404":
          description: Юзер не найден
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
        "401":
          description: Запрос не авторизован
      security:
        - bearerAuth: [ ]
  /roles/get/roles/{role}:
    get:
      tags:
        - Роли/должности
      summary: Получить все роли (связки юзера и роли в БД) по имени енама роли
      description: Возвращает список ролей по имени енама роли
      operationId: getRolesByRole
      parameters:
        - name: role
          in: path
          description: Роль (имя енама)
          required: true
          schema:
            type: string
            enum:
              - ADMINISTRATION
              - HOSTEL_SUPERVISOR
              - FLOOR_SUPERVISOR
              - RESPONSIBLE_KITCHEN
              - RESPONSIBLE_HALL
              - RESPONSIBLE_GYM
              - WORKER_GYM
              - RESPONSIBLE_FIRE_SAFETY
              - RESPONSIBLE_SANITARY
              - RESPONSIBLE_INTERNET
              - RESPONSIBLE_SOOP
              - WORKER_FIRE_SAFETY
              - WORKER_SANITARY
              - WORKER_SOOP
              - STUDENT
      responses:
        "200":
          description: Успешное получение ролей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleResponseDto"
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
        "401":
          description: Запрос не авторизован
      security:
        - bearerAuth: [ ]
  /roles/{userId}:
    delete:
      tags:
        - Роли/должности
      summary: Удалить роль
      description: Удаляет роль с юзера (снимает с должности) по ID юзера и имени
        енама роли
      operationId: deleteRole
      parameters:
        - name: userId
          in: path
          description: ID юзера
          required: true
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          description: Имя енама роли
          required: true
          schema:
            type: string
            enum:
              - ADMINISTRATION
              - HOSTEL_SUPERVISOR
              - FLOOR_SUPERVISOR
              - RESPONSIBLE_KITCHEN
              - RESPONSIBLE_HALL
              - RESPONSIBLE_GYM
              - WORKER_GYM
              - RESPONSIBLE_FIRE_SAFETY
              - RESPONSIBLE_SANITARY
              - RESPONSIBLE_INTERNET
              - RESPONSIBLE_SOOP
              - WORKER_FIRE_SAFETY
              - WORKER_SANITARY
              - WORKER_SOOP
              - STUDENT
      responses:
        "204":
          description: Успешное удаление роли
        "403":
          description: Нет прав управлять ролью
        "404":
          description: Не найден юзер
        "409":
          description: Юзер не назначен на роль
        "507":
          description: Не удалось удалить роль
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
        "401":
          description: Запрос не авторизован
      security:
        - bearerAuth: [ ]
components:
  schemas:
...
    RoleSetDto:
      type: object
      description: Тело запроса присваивания роли юзеру
      properties:
        user:
          type: string
          format: uuid
          description: "ID юзера, которому присваивается роль"
        role:
          type: string
          description: Присваиваемая роль в виде имени енама
          enum:
            - ADMINISTRATION
            - HOSTEL_SUPERVISOR
            - FLOOR_SUPERVISOR
            - RESPONSIBLE_KITCHEN
            - RESPONSIBLE_HALL
            - RESPONSIBLE_GYM
            - WORKER_GYM
            - RESPONSIBLE_FIRE_SAFETY
            - RESPONSIBLE_SANITARY
            - RESPONSIBLE_INTERNET
            - RESPONSIBLE_SOOP
            - WORKER_FIRE_SAFETY
            - WORKER_SANITARY
            - WORKER_SOOP
            - STUDENT
      required:
        - role
        - user
    RoleResponseDto:
      type: object
      description: Ответ с информацией о роли (связки юзера и роли в БД)
      properties:
        id:
          type: string
          format: uuid
          description: ID роли (связки юзера и роли в БД)
        user:
          type: string
          format: uuid
          description: "ID юзера, которому присвоили роль"
        role:
          type: string
          description: "Название роли, которую присвоили юзеру"
          example: Студент
...
    RoleEditDto:
      type: object
      description: Тело запроса изменения роли юзера
      properties:
        id:
          type: string
          format: uuid
          description: "ID роли (связки юзера и роли в БД), которую нужно изменить"
        role:
          type: string
          description: "Имя енама роли, на которую происходит изменение"
          enum:
            - ADMINISTRATION
            - HOSTEL_SUPERVISOR
            - FLOOR_SUPERVISOR
            - RESPONSIBLE_KITCHEN
            - RESPONSIBLE_HALL
            - RESPONSIBLE_GYM
            - WORKER_GYM
            - RESPONSIBLE_FIRE_SAFETY
            - RESPONSIBLE_SANITARY
            - RESPONSIBLE_INTERNET
            - RESPONSIBLE_SOOP
            - WORKER_FIRE_SAFETY
            - WORKER_SANITARY
            - WORKER_SOOP
            - STUDENT
      required:
        - id
        - role
    UserResponseWithRoleDto:
      type: object
      description: Ответ со всей инфой о юзере вместе с ролями
      properties:
        id:
          type: string
          format: uuid
          description: ID
        firstName:
          type: string
          description: Имя
        lastName:
          type: string
          description: Фамилия
        middleName:
          type: string
          description: Отчество
        email:
          type: string
          description: Имейл
        phoneNumber:
          type: string
          description: Номер телефона
        roomNumber:
          type: string
          description: Номер комнаты
        roles:
          type: array
          description: Список ролей
          items:
            type: string
        tgLink:
          type: string
          description: Имя в ТГ
        vkLink:
          type: string
          description: Имя в ВК
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT